requirements.txt
django_rest_framework_mongoengine==3.3.1
Django==1.11.2
mongoengine==0.13.0
